\name{gryphons}
\docType{data}
\alias{gryphons}
\title{Example dataset for pedantics examples and titorial}
\description{
  This contains pedigree and life history data of a fictional
population.  The data are relevent to power and sensitivity 
analyses for quantitative genetic studies of natural populations.
}
\usage{gryphons}
\format{A table.}
\keyword{datasets}

