\name{pedantics-package}
\alias{pedantics-package}
\alias{pedantics}
\docType{package}
\title{
Tools to facilitate quantitative genetic studies of natural populations, especially with respect to the use of pedigrees in such problems.
}
\description{
\code{pedantix} contains three types of functions.  The first are functions specifically designed to aid power and sensitivity analyses for quantitative genetic studies, particulary with thought to accommodating the proglems and data structures that arise in data from natural populations.  There are basic utility functions for manipulating pedigrees.  Finally there are functions for visualizing and statistically characterizing pedigrees.
}
\details{
\tabular{ll}{
Package: \tab pedantics\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2009-07-21\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}
See the tutorial, pedantics-Tutorial.pdf for detailed example analyses using \code{pedantics}
}
\author{
Michael Morrissey \email{michael.morrissey@ed.ac.uk}

Maintainer: Michael Morrissey \email{michael.morrissey@ed.ac.uk}
}
\references{Morrissey et al. 2007. Journal of Evolutionary Biology 20:2309-2321., Morrissey, M.B, and A.J. Wilson, 2009. pedantics, an R package for pedigree-based genetic simulation, and pedigree manipulation, characterisation, and viewing. Molecular Ecology Resources.}

\keyword{ package }

